JA: Fixed son class. (CHANGE SLEEP 5 IN EXEC MAKEFILE LATER). Changed ATT to uint16. Changed sound to PWM. Changed control structure for DBT. Added DEBUG_PRINTs for tp9

+++ tp/tp9/lib/robot.cpp
@@ -41,8 +41,8 @@ void Robot::arreterSon() {
     son_.arreter();
 }
 
-void Robot::attendre(uint8_t multiplicateur) {
-    for (uint8_t i = 0; i < multiplicateur; ++i) {
+void Robot::attendre(uint16_t multiplicateur) {
+    for (uint16_t i = 0; i < multiplicateur; ++i) {
         _delay_ms(25); // Delay of 25 ms per unit of multiplicateur
     }
 }
